rename files and folders. remove deadcode
