+++ src/components/GinContainer/index.tsx
@@ -1,5 +1,5 @@
-import React, { FC } from "react";
 import "./styles.css";
+import { FC } from "react";
 import ginImage from "../../images/ginSour.jpeg";
 import { Link } from "react-router-dom";
 
@@ -8,7 +8,11 @@ const GinContainer: FC = () => {
     <div>
       <Link to="/GinDeck">
         <div className="ginButton">
-          <img className="ginImage" src={ginImage} />
+          <img
+            className="ginImage"
+            src={ginImage}
+            alt="Gin logo to act as a button for its category"
+          />
           <h3>20 Gins that will get you loose</h3>
         </div>
       </Link>
