+++ src/components/GinDeck/index.tsx
@@ -1,5 +1,5 @@
 import "./styles.css";
-import Error from "../Error/error";
+import Error from "../error";
 import { FC, useState, useEffect } from "react";
 import { fetchSpecialtyCocktails } from "../../apiCalls";
 import { Drink } from "../Types";
@@ -7,12 +7,16 @@ import { Link } from "react-router-dom";
 
 const GinDeck: FC = () => {
   const [drinkList, setDrinkList] = useState([]);
-  const [error, setError] = useState(false)
+  const [error, setError] = useState(false);
 
   useEffect(() => {
     fetchSpecialtyCocktails("search.php?s=gin")
-      .then((data) => {setDrinkList(data.drinks)})
-      .catch(() => {setError( true )});
+      .then((data) => {
+        setDrinkList(data.drinks);
+      })
+      .catch(() => {
+        setError(true);
+      });
   }, []);
 
   const formatGins = () => {
@@ -29,11 +33,11 @@ const GinDeck: FC = () => {
       );
     });
   };
-  
-    return (
-      <div className="ginContainer">
-        {error ? <Error /> : <>{formatGins()}</>}
-      </div>
-    );
-}
+
+  return (
+    <div className="ginContainer">
+      {error ? <Error /> : <>{formatGins()}</>}
+    </div>
+  );
+};
 export default GinDeck;
