--- src/components/MartiniDeck/MartiniDeck.css
@@ -1,45 +1,0 @@
-.martiniContainer {
-display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    grid-template-rows: repeat(4, 1fr);
-    grid-column-gap: 5px;
-    grid-row-gap: 5px;
-}
-
-.deckContainer {
-    height: 65%;
-    width:65%;
-    margin-top: 40px;
-    margin-bottom: 40px;
-    margin-left: 20%;
-    margin-right: 20%;
-    align-items: center;
-    display: flex;
-    align-items: center;
-    flex-direction: row;
-}
-
-.drink-image {
-    display: flex;
-    flex-direction: column;
-    height: 80%;
-    width: 80%;
-    margin-top: 10%;
-}
-
-.martiniDrink {
-    color: black;
-    background-color: azure;
-    display:flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: baseline;
-    font-family: 'Montserrat', sans-serif;
-}
-
-.martiniDrink:hover {
-    cursor: pointer;
-    transform: scale(1.1);
-    transition: all .8s;
-    box-shadow: 15px 15px 10px rgba(0, 0, 0, 0.3);
-}
