+++ src/components/Types.ts
@@ -15,11 +15,6 @@ export type Drink = {
   strIngredient4?: string | null;
   strIngredient5?: string | null;
   strIngredient6?: string | null;
-  strIngredient7?: string | null;
-  strIngredient8?: string | null;
-  strIngredient9?: string | null;
-  strIngredient10?: string | null;
-  strIngredient11?: string | null;
   error?: false | true;
 };
