+++ src/components/app/index.tsx
@@ -1,11 +1,11 @@
-import "./App.css";
-import NavBar from "../NavBar/NavBar";
-import GinDeck from "../GinDeck";
-import MartiniDeck from "../MartiniDeck/MartiniDeck";
-import AmarettoDeck from "../AmarettoDeck";
-import DetailCards from "../DetailCards";
-import AppContainer from "../AppContainer/AppContainer";
-import Error from "../Error/error";
+import "./styles.css";
+import NavBar from "../navBar";
+import GinDeck from "../ginDeck";
+import MartiniDeck from "../martiniDeck";
+import AmarettoDeck from "../amarettoDeck";
+import DetailCards from "../detailCards";
+import AppContainer from "../appContainer";
+import Error from "../error";
 import { FC, useState, useEffect } from "react";
 import { fetchAllCocktails } from "../../apiCalls";
 import { Route } from "react-router-dom";
